refactor: extract common.ts logic into separate files
