+++ src/components/Integration.tsx
@@ -11,8 +11,8 @@ import WeekSchedule from "@/components/schedule/WeekSchedule";
 import AppBar from "@/components/utils/AppBar";
 import Logo from "@/components/utils/Logo";
 import PageHead from "@/components/utils/PageHead";
+import { classConfigRecurrentId } from "@/lib/helpers/recurrentId";
 import { useUserConfig } from "@/lib/hooks/useUserConfig";
-import { classConfigRecurrentId } from "@/lib/integrations/common";
 import {
     ClassPopularityIndex,
     IntegrationProfile,
