+++ src/components/modals/ClassInfo/ClassInfoModal.tsx
@@ -2,7 +2,7 @@ import { Modal } from "@mui/material";
 import React, { Dispatch, SetStateAction } from "react";
 
 import ClassInfo from "@/components/modals/ClassInfo/ClassInfo";
-import { classRecurrentId } from "@/lib/integrations/common";
+import { classRecurrentId } from "@/lib/helpers/recurrentId";
 import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex, RezervoClass } from "@/types/rezervo";
 
 const ClassInfoModal = ({
