+++ src/components/schedule/class/ClassCard.tsx
@@ -6,8 +6,8 @@ import React, { useEffect, useState } from "react";
 import ClassPopularityMeter from "@/components/schedule/class/ClassPopularityMeter";
 import ClassUserAvatar from "@/components/schedule/class/ClassUserAvatar";
 import RippleBadge from "@/components/utils/RippleBadge";
+import { isClassInThePast } from "@/lib/helpers/date";
 import { useUserSessions } from "@/lib/hooks/useUserSessions";
-import { isClassInThePast } from "@/lib/integrations/common";
 import { randomElementFromArray } from "@/lib/utils/arrayUtils";
 import { hexColorHash, hexWithOpacityToRgb } from "@/lib/utils/colorUtils";
 import { EnterLeaveAnimation, OVER_THE_TOP_ANIMATIONS } from "@/types/animation";
