+++ src/components/schedule/class/ClassPopularityMeter.tsx
@@ -3,8 +3,8 @@ import SpeedRoundedIcon from "@mui/icons-material/SpeedRounded";
 import { Tooltip } from "@mui/material";
 
 import RippleBadge from "@/components/utils/RippleBadge";
-import { isClassInThePast } from "@/lib/integrations/common";
-import { determineClassPopularity, stringifyClassPopularity } from "@/lib/popularity";
+import { isClassInThePast } from "@/lib/helpers/date";
+import { determineClassPopularity, stringifyClassPopularity } from "@/lib/helpers/popularity";
 import { ClassPopularity, RezervoClass, StatusColors } from "@/types/rezervo";
 
 const ClassPopularityIcon = ({ popularity }: { popularity: ClassPopularity }) => {
