+++ src/lib/consts.ts
@@ -1,5 +1,3 @@
-export const GROUP_BOOKING_URL = "https://www.sit.no/trening/gruppe";
-
 export const TIME_ZONE = "Europe/Oslo";
 export const LOCALE = "no";
