+++ src/lib/integrations/fsc/adapters.ts
@@ -1,9 +1,9 @@
 import { DateTime } from "luxon";
 
+import determineActivityCategory from "@/lib/helpers/activityCategorization";
+import { getDateTime, zeroIndexedWeekday } from "@/lib/helpers/date";
 import { IntegrationIdentifier } from "@/lib/integrations/active";
-import { determineActivityCategory } from "@/lib/integrations/common";
 import { DetailedFscClass, DetailedFscWeekSchedule } from "@/lib/integrations/fsc/types";
-import { getDateTime, zeroIndexedWeekday } from "@/lib/utils/dateUtils";
 import { RezervoClass, RezervoWeekSchedule } from "@/types/rezervo";
 
 function fscToRezervoClass(fscClass: DetailedFscClass): RezervoClass {
