+++ src/lib/integrations/fsc/fetchers.ts
@@ -1,6 +1,7 @@
 import { DateTime } from "luxon";
 
 import { TIME_ZONE } from "@/lib/consts";
+import { calculateMondayOffset } from "@/lib/helpers/date";
 import {
     DetailedFscClass,
     DetailedFscWeekSchedule,
@@ -10,7 +11,6 @@ import {
     FscWeekSchedule,
     FscWeekScheduleResponse,
 } from "@/lib/integrations/fsc/types";
-import { calculateMondayOffset } from "@/lib/utils/dateUtils";
 
 function fscWeekScheduleUrl(fromDate: DateTime) {
     return `https://fsc.no/api/v1/businessunits/8/groupactivities?period_start=${fromDate.toUTC()}&period_end=${fromDate
