+++ src/lib/integrations/sit/adapters.ts
@@ -1,6 +1,6 @@
+import { getDateTime } from "@/lib/helpers/date";
 import { IntegrationIdentifier } from "@/lib/integrations/active";
 import { SitClass, SitDaySchedule, SitWeekSchedule } from "@/lib/integrations/sit/types";
-import { getDateTime } from "@/lib/utils/dateUtils";
 import { RezervoClass, RezervoDaySchedule, RezervoWeekSchedule } from "@/types/rezervo";
 
 function sitToRezervoClass(sitClass: SitClass): RezervoClass {
