+++ src/lib/integrations/sit/fetchers.ts
@@ -1,8 +1,8 @@
 import { DateTime } from "luxon";
 
-import { GROUP_BOOKING_URL, TIME_ZONE } from "@/lib/consts";
+import { TIME_ZONE } from "@/lib/consts";
+import { calculateMondayOffset } from "@/lib/helpers/date";
 import { SitDaySchedule, SitWeekSchedule } from "@/lib/integrations/sit/types";
-import { calculateMondayOffset } from "@/lib/utils/dateUtils";
 
 function sitScheduleUrl(token: string, fromISO: string | null = null) {
     return (
@@ -12,8 +12,12 @@ function sitScheduleUrl(token: string, fromISO: string | null = null) {
     );
 }
 
+function sitGroupBookingUrl() {
+    return "https://www.sit.no/trening/gruppe";
+}
+
 function fetchSitPublicToken() {
-    return fetch(GROUP_BOOKING_URL)
+    return fetch(sitGroupBookingUrl())
         .then((res) => res.text())
         .then((text) => text.replace(/[\n\r]/g, "").replace(/\s+/g, " "))
         .then((soup) => {
