+++ src/pages/api/schedule.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 
-import { fetchRezervoWeekSchedule } from "@/lib/integrations/common";
+import { fetchRezervoWeekSchedule } from "@/lib/helpers/fetchers";
 import { sitToRezervoWeekSchedule } from "@/lib/integrations/sit/adapters";
 import { fetchSitWeekSchedule } from "@/lib/integrations/sit/fetchers";
 import { serializeWeekSchedule } from "@/lib/serialization/serializers";
